Add numbers to tabs; revert font weight yikes
