+++ src/index.css
@@ -1,7 +1,6 @@
 body {
   margin: 0;
   font-family: 'Source Sans Pro', sans-serif;
-  font-weight: 600;
   color: rgba(0, 0, 0, 0.85);
   -webkit-font-smoothing: antialiased;
 }
